Fix: Fully synchronized TeamDao and MainActivity to use Long IDs

+++ android-app/app/src/main/java/com/duallive/app/data/dao/TeamDao.kt
@@ -7,7 +7,7 @@ import kotlinx.coroutines.flow.Flow
 @Dao
 interface TeamDao {
     @Query("SELECT * FROM teams WHERE leagueId = :leagueId")
-    fun getTeamsByLeague(leagueId: Int): Flow<List<Team>>
+    fun getTeamsByLeague(leagueId: Long): Flow<List<Team>>
 
     @Insert(onConflict = OnConflictStrategy.REPLACE)
     suspend fun insertTeam(team: Team)
